project-orc/requirements.txt
daytona-sdk>=0.10.0
python-dotenv>=1.0.0
pytest>=8.0.0
